attempt to read the dirs
